hw4

+++ hw4/homework/train.py
@@ -103,7 +103,7 @@ def train(args):
             logit = model(img)
             print(logit)
             print(label)
-            loss_val = loss(logit.long(), label.long())
+            loss_val = loss(logit.float(), label.float())
             if train_logger is not None and global_step % 100 == 0:
                 train_logger.add_image('image', img[0], global_step)
                 train_logger.add_image('label', np.array(dense_transforms.label_to_pil_image(label[0].cpu()).
